drfproj/requirements.txt
aenum==3.1.15
ajax==3.0.0
alabaster==0.7.13
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph==0.17.3
anyio==4.2.0
appdirs==1.4.3
Appium-Python-Client==3.1.1
apturl==0.5.2
asgiref==3.7.2
assertpy==1.1
astroid==2.15.8
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.33.0
Babel==2.12.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
black==23.1.0
blinker==1.6.2
Bootstrap-Flask==2.3.0
bootstrap-py==1.0.2
bottle==0.12.25
bottle-websocket==0.2.9
Brlapi==0.7.0
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chromedriver==2.24.1
chromedriver-py==114.0.5735.90
click==8.1.7
colorama==0.4.3
colorthief==0.2.1
command-not-found==0.3
ConfigArgParse==1.7
contourpy==1.0.7
cprint==1.2.2
cryptography==41.0.3
cupshelpers==1.0
curlify==2.2.1
cycler==0.11.0
darkdetect==0.8.0
dbus-python==1.2.16
ddt==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
delayed-assert==0.3.6
dill==0.3.8
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.13
django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot.git@a840496d183df89965eeb33c9a5dd9ea3919dfff
django-appconf==1.0.6
django-ipware==7.0.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
docutils==0.20.1
dominate==2.8.0
duplicity==0.8.12.0
EasyProcess==1.1
Eel==0.16.0
entrypoint2==1.1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
exrex==0.11.0
Faker==22.7.0
fast-map==0.1.0
fastapi==0.109.2
fastapi-utils==0.2.1
fasteners==0.14.1
filelock==3.0.12
flake8==6.0.0
Flask==2.3.2
Flask-API==3.1
Flask-BasicAuth==0.2.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-CKEditor==0.4.6
Flask-Cors==4.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.39.4
future==0.18.2
geckodriver-autoinstaller==0.1.0
get-chrome-driver==1.3.10
GetPixelColor==0.2.19
gevent==21.12.0
gevent-websocket==0.10.1
geventhttpclient==2.0.11
gitdb==4.0.10
GitPython==3.1.34
greenlet==1.1.2
gTTS==2.5.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.14.0
hyperframe==6.0.1
idna==2.8
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
influxdb-client==1.40.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
jsonpath-ng==1.6.1
kaitaistruct==0.10
keyring==18.0.1
Kivy==2.2.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.10.0
line-profiler==4.1.2
lockfile==0.12.2
locust==2.12.1
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mpg123==0.4
msgpack==1.0.7
mss==9.0.1
mypy-extensions==1.0.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.8.0.76
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.5.0
parameterized==0.9.0
paramiko==2.6.0
pathspec==0.12.1
pbr==5.11.1
pdfminer.six==20200517
pdfplumber==0.5.24
pexpect==4.6.0
pguard==0.4.0
pgzero==1.2.1
pillow==10.3.0
platformdirs==4.2.0
playsound==1.3.0
pluggy==1.4.0
ply==3.11
prometheus-client==0.19.0
protobuf==3.6.1
psutil==5.9.8
psycogreen==1.0.2
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
PyAutoGUI==0.9.54
PyByteBuffer==1.0.5
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pycups==1.9.73
pydantic==1.10.14
pydub==0.25.1
pyflakes==3.0.1
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.36.0
pyinstaller==5.10.0
pyinstaller-hooks-contrib==2023.2
PyJWT==2.8.0
pylint==2.17.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
pymssql==2.2.7
PyNaCl==1.3.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.4.4
pytest-html==3.2.0
pytest-json-report==1.5.0
pytest-metadata==2.0.4
pytest-order==1.2.0
pytest-ordering==0.6
pytest-reportportal==5.1.5
pytest-rerunfailures==13.0
pytest-xdist==3.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-ipware==3.0.0
python3-xlib==0.15
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
PyVirtualDisplay==3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
reactivex==4.0.4
redis==4.5.5
remi==2022.7.27
reportlab==4.2.0
reportportal-client==5.2.5
requests==2.31.0
requests-unixsocket==0.2.0
roundrobin==0.0.4
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
SecretStorage==2.3.1
selenium==4.18.1
selenium-wire==5.1.0
simple-term-menu==1.6.1
simplejson==3.16.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spotipy==2.23.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.36.3
system-service==0.3
systemd-python==234
taskgroup==0.0.0a4
termcolor==2.2.0
testresources==2.0.1
threadpoolctl==3.2.0
tkcalendar==1.6.1
tkdesigner==1.0.7
tkinter-tooltip==2.1.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
ttkbootstrap==1.10.1
typer==0.9.0
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.27.0.post1
virtualenv==20.0.17
virtualenvwrapper-win==1.2.7
visitor==0.1.3
wadllib==1.3.3
Wand==0.6.11
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.4.2
Werkzeug==2.3.7
whichcraft==0.6.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.0.1
xkit==0.0.0
youtube-dl==2021.12.17
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
